test_files_add2

+++ symtable.cpp
@@ -12,7 +12,7 @@ void SymbolTable::cleanStack(){
 
 int SymbolTable::getAdress(string ID){
         int adress = 0;
-        for (int i=0; i<vectorOfSymbols.size(); i++) {
+        for (long unsigned int i=0; i<vectorOfSymbols.size(); i++) {
             if (ID == vectorOfSymbols[i].name) {
                 break;
             }
@@ -22,7 +22,7 @@ int SymbolTable::getAdress(string ID){
 }
 
 void SymbolTable::printSymbolTable(){
-    for (int i=0; i<vectorOfSymbols.size(); i++){
+    for (long unsigned int i=0; i<vectorOfSymbols.size(); i++){
         cout << vectorOfSymbols[i].name << std::setw(6) << vectorOfSymbols[i].size << endl;
     }
 }
